Backwards Support Not Quite Right
Tried to add in support to go backwards.  Forwards support still seems to work in GNU\Linux but wasted a bunch of time this evening trying to troubleshoot some weird bug with a !! prelude index too big crash on MS Windows.  I don't think that's my code I think there maybe something specific to the MS Windows Libraries that was causing that weirdness.  I won't be able to get back to this for a few days because I have personal things I need to take care of besides my day job.